# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/nutri/day-types/{dayTypeId}/meals": {
      "post": {
        "description": "Corpo: {name, position, horario?}. `position` é ÚNICA no tipo-de-dia e é a chave que pareia refeições entre tipos-de-dia (a troca de tipo-de-dia depende dela) — duplicar responde 409. `horario` é informativo (HH:MM ou HH:MM:SS): não dirige \"o agora\".",
        "operationId": "RefeicaoController_criarRefeicao",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dayTypeId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Almoço",
                  "position": 2,
                  "horario": "12:30"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "a refeição criada, sem opções"
          },
          "400": {
            "description": "name/position/horario inválidos"
          },
          "403": {
            "description": "x-nutri-key ausente/errada"
          },
          "404": {
            "description": "tipo-de-dia não encontrado"
          },
          "409": {
            "description": "position já usada neste tipo-de-dia"
          }
        },
        "summary": "Criar uma refeição (slot) no tipo-de-dia",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
    "/nutri/meals/{mealId}": {
      "patch": {
        "description": "PATCH parcial de {name, position, horario}. `horario: null` limpa. Mover para uma position ocupada responde 409.",
        "operationId": "RefeicaoController_atualizarRefeicao",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a refeição depois da alteração"
          },
          "404": {
            "description": "refeição não encontrada"
          },
          "409": {
            "description": "position já usada neste tipo-de-dia"
          }
        },
        "summary": "Editar a refeição",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      },
      "delete": {
        "description": "RECUSA com 409 se houver registro nesta refeição.",
        "operationId": "RefeicaoController_excluirRefeicao",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "refeição excluída"
          },
          "404": {
            "description": "refeição não encontrada"
          },
          "409": {
            "description": "há registro nesta refeição"
          }
        },
        "summary": "Excluir a refeição (com opções e itens)",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
    "/nutri/meals/{mealId}/options": {
      "post": {
        "description": "Corpo: {label, isDefault?}. A PRIMEIRA opção da refeição nasce padrão mesmo sem pedir (refeição com opções e nenhuma padrão não tem o que mostrar). Criar com isDefault:true desmarca as irmãs — exatamente uma padrão por refeição.",
        "operationId": "RefeicaoController_criarOpcao",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mealId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "label": "Arroz e carne",
                  "isDefault": true
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "a opção criada"
          },
          "400": {
            "description": "label ausente/vazio"
          },
          "404": {
            "description": "refeição não encontrada"
          }
        },
        "summary": "Criar uma opção da refeição (os \"3 almoços\")",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
    "/nutri/options/{optionId}": {
      "patch": {
        "description": "PATCH parcial de {label, isDefault}. `isDefault: true` promove esta e desmarca as irmãs. `isDefault: false` na única padrão responde 409 — marque outra como padrão em vez de desmarcar esta.",
        "operationId": "RefeicaoController_atualizarOpcao",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "a opção depois da alteração"
          },
          "404": {
            "description": "opção não encontrada"
          },
          "409": {
            "description": "tentou desmarcar a única opção padrão"
          }
        },
        "summary": "Editar a opção",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      },
      "delete": {
        "description": "RECUSA com 409 se for a ÚNICA opção da refeição, ou se algum registro aponta para ela. Se era a padrão havendo outras, promove outra no mesmo ato.",
        "operationId": "RefeicaoController_excluirOpcao",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "opção excluída"
          },
          "404": {
            "description": "opção não encontrada"
          },
          "409": {
            "description": "única opção, ou apontada por registro"
          }
        },
        "summary": "Excluir a opção (com os itens)",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
    "/nutri/options/{optionId}/items": {
      "post": {
        "description": "Corpo: {foodId, quantityGrams, isLocked?, substitutionGroupId?}. `isLocked` e `substitutionGroupId` são a marcação de flexibilidade inteira e são MUTUAMENTE EXCLUSIVOS (travado não troca) — os dois juntos respondem 400. O alimento precisa participar do grupo informado, senão 422: é o vínculo que carrega a porção de referência, sem a qual a troca não sabe reescalar a quantidade.",
        "operationId": "RefeicaoController_criarItem",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optionId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "foodId": "…",
                  "quantityGrams": 120,
                  "isLocked": false,
                  "substitutionGroupId": "…"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "o item criado"
          },
          "400": {
            "description": "gramas ≤ 0, ou isLocked junto com substitutionGroupId"
          },
          "404": {
            "description": "opção, alimento ou grupo não encontrado"
          },
          "422": {
            "description": "o alimento não participa do grupo informado"
          }
        },
        "summary": "Adicionar um item à opção",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
    "/nutri/items/{itemId}": {
      "patch": {
        "description": "PATCH parcial de {foodId, quantityGrams, isLocked, substitutionGroupId}. A marcação de flexibilidade é avaliada em CONJUNTO com o que já está gravado: mandar só `isLocked: true` num item que tem grupo responde 400, não \"resolve\" por precedência.",
        "operationId": "RefeicaoController_atualizarItem",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "o item depois da alteração"
          },
          "400": {
            "description": "valores inválidos ou contraditórios"
          },
          "404": {
            "description": "item, alimento ou grupo não encontrado"
          },
          "422": {
            "description": "o alimento não participa do grupo informado"
          }
        },
        "summary": "Editar o item",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      },
      "delete": {
        "description": "Sem bloqueador: nada referencia `meal_item` (o snapshot do registro aponta para `food`). Sai direto.",
        "operationId": "RefeicaoController_excluirItem",
        "parameters": [
          {
            "name": "x-nutri-key",
            "in": "header",
            "description": "credencial stub da nutri (env NUTRI_API_KEY; fail-closed)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "item excluído"
          },
          "404": {
            "description": "item não encontrado"
          }
        },
        "summary": "Excluir o item",
        "tags": [
          "Editor de plano (só nutri)"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bamboo API",
    "description": "API da alça do paciente — ver \"o agora\" (US1) e substituir um alimento dentro do grupo, com quantidade recalculada e medida caseira (US2). v0: auth stub (paciente fixo por env).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}